backend/data_management_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
minio==7.2.5
python-jose[cryptography]==3.3.0
h5py==3.10.0
